feat: Details renders comments
