+++ src/view/film-details.js
@@ -1,12 +1,6 @@
-import {formatCommentDataTime, formatDate, formatTime} from '../utils/date-time';
+import {formatDate, formatTime} from '../utils/date-time';
 import SmartView from './smart-view';
 
-const EMOJIS_PATHS = {
-  'emoji-smile': './images/emoji/smile.png',
-  'emoji-sleeping': './images/emoji/sleeping.png',
-  'emoji-puke': './images/emoji/puke.png',
-  'emoji-angry': './images/emoji/angry.png',
-};
 
 const createGenresListTemplate = (genres) =>
   genres.map((genre) => (
@@ -14,27 +8,6 @@ const createGenresListTemplate = (genres) =>
       <span class="film-details__genre">${genre}</span>
     `
   )).join(' ');
-const getEmotionPath = (emotion) =>  `./images/emoji/${emotion}.png`;
-
-const createCommentsList = (comments) =>
-  comments.map((comment) =>
-    `
-      <li class="film-details__comment">
-      <span class="film-details__comment-emoji">
-        <img src="${getEmotionPath(comment.emotion)}" width="55" height="55" alt="emoji-smile">
-      </span>
-      <div>
-        <p class="film-details__comment-text">${comment.comment}</p>
-        <p class="film-details__comment-info">
-          <span class="film-details__comment-author">${comment.author}</span>
-          <span class="film-details__comment-day">${formatCommentDataTime(comment.date)}</span>
-          <button class="film-details__comment-delete">Delete</button>
-        </p>
-      </div>
-    </li>
-  `
-  ).join('');
-
 
 const commaSeparatedList = (phrases) => phrases.join(', ');
 const getRelevantActiveClass = (isActiveFlag) => isActiveFlag ? 'film-details__control-button--active' : '';
@@ -57,7 +30,6 @@ const createFilmDetailsTemplate = (film) => {
     watchlist,
     watched,
     favorite,
-    comments
   } = film;
 
   return (
@@ -134,42 +106,6 @@ const createFilmDetailsTemplate = (film) => {
 
     <div class="film-details__bottom-container">
       <section class="film-details__comments-wrap">
-        <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${comments.length}</span></h3>
-
-        <ul class="film-details__comments-list">
-          ${createCommentsList(comments)}
-        </ul>
-
-        <div class="film-details__new-comment">
-          <div class="film-details__add-emoji-label">
-          </div>
-          <input hidden name="selected-emoji"> 
-          <label class="film-details__comment-label">
-            <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-          </label>
-
-          <div class="film-details__emoji-list">
-            <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile">
-            <label class="film-details__emoji-label" for="emoji-smile">
-              <img src=${EMOJIS_PATHS['emoji-smile']} width="30" height="30" alt="emoji">
-            </label>
-
-            <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
-            <label class="film-details__emoji-label" for="emoji-sleeping">
-              <img src=${EMOJIS_PATHS['emoji-sleeping']} width="30" height="30" alt="emoji">
-            </label>
-
-            <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
-            <label class="film-details__emoji-label" for="emoji-puke">
-              <img src=${EMOJIS_PATHS['emoji-puke']} width="30" height="30" alt="emoji">
-            </label>
-
-            <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
-            <label class="film-details__emoji-label" for="emoji-angry">
-              <img src=${EMOJIS_PATHS['emoji-angry']} width="30" height="30" alt="emoji">
-            </label>
-          </div>
-        </div>
       </section>
     </div>
   </form>
@@ -180,13 +116,13 @@ const createFilmDetailsTemplate = (film) => {
 export default class FilmDetails extends SmartView{
   #film = null;
   #closeButton = null;
-  #emojis = [];
+  #commentsContainer = null;
 
   constructor(film, externalHandlers) {
     super();
     this.#film = film;
     this.#closeButton = this.element.querySelector('.film-details__close-btn');
-    this.#emojis = Array.from(this.element.querySelectorAll('.film-details__emoji-label'));
+    this.#commentsContainer = this.element.querySelector('.film-details__bottom-container');
 
     this._externalHandlers.closeDetails = externalHandlers.closeDetailsHandler;
     this._externalHandlers.clickWatchList = externalHandlers.clickWatchListHandler(this.#film);
@@ -203,11 +139,14 @@ export default class FilmDetails extends SmartView{
 
     this.#closeButton.addEventListener('click', this.#clickCloseHandler);
     document.addEventListener('keydown', this.#onKeyDownHandler);
-    this.#emojis.forEach((emojiLabel) => emojiLabel.addEventListener('click', this.#clickEmotion));
   }
 
-  get filmID() {
+  get filmId() {
     return this.#film.id;
+  }
+
+  get commentsContainer() {
+    return this.#commentsContainer;
   }
 
   #clickWatchList = (event) => {
@@ -225,23 +164,6 @@ export default class FilmDetails extends SmartView{
     this._externalHandlers.clickFavorite(this.#film);
   }
 
-  #clickEmotion = (event) => {
-    const emoji = event.currentTarget.attributes.for.value;
-    const emojisPlace = this.element.querySelector('.film-details__add-emoji-label');
-    const prevEmotion = emojisPlace.firstChild;
-
-    const newEmotion = document.createElement('img');
-    newEmotion.src = EMOJIS_PATHS[emoji];
-    newEmotion.height=55;
-    newEmotion.width=55;
-    newEmotion.alt='emoji';
-
-    if (prevEmotion) {
-      emojisPlace.removeChild(prevEmotion);
-    }
-    emojisPlace.appendChild(newEmotion);
-  }
-
   #clickCloseHandler = (event) => {
     event.preventDefault();
     this._externalHandlers.closeDetails();
@@ -250,7 +172,7 @@ export default class FilmDetails extends SmartView{
   #onKeyDownHandler = (event) => {
     if (event.code === 'Enter' && (event.ctrlKey || event.metaKey)) {
       event.preventDefault();
-      //TO-DO this._externalHandlers.submitDetails();
+      //TODO this._externalHandlers.submitDetails();
     }
     if (event.key === 'Escape' || event.key === 'Esc') {
       event.preventDefault();
